Turn static __open_unix_server(), used only once
