+++ main/gridinit.c
@@ -848,6 +848,57 @@ servers_save_fd(int fd, const char *url)
 	return TRUE;
 }
 
+static int
+__open_unix_server(const char *path)
+{
+	struct sockaddr_un local = {};
+
+	if (!path || strlen(path) >= sizeof(local.sun_path)) {
+		errno = EINVAL;
+		return -1;
+	}
+
+	/* Create ressources to monitor */
+#ifdef SOCK_CLOEXEC
+# define SOCK_FLAGS SOCK_CLOEXEC
+#else
+# define SOCK_FLAGS 0
+#endif
+
+	int sock = socket(PF_UNIX, SOCK_STREAM | SOCK_FLAGS, 0);
+	if (sock < 0)
+		return -1;
+
+#ifndef SOCK_CLOEXEC
+# ifdef FD_CLOEXEC
+	(void) fcntl(sock, F_SETFD, fcntl(sock, F_GETFD)|FD_CLOEXEC);
+# endif
+#endif
+
+	/* Bind to file */
+	local.sun_family = AF_UNIX;
+	g_strlcpy(local.sun_path, path, sizeof(local.sun_path)-1);
+
+	if (-1 == bind(sock, (struct sockaddr *)&local, sizeof(local)))
+		goto label_error;
+
+	/* Listen on that socket */
+	if (-1 == listen(sock, 65536))
+		goto label_error;
+
+	errno = 0;
+	return sock;
+
+label_error:
+	if (sock >= 0) {
+		typeof(errno) errsav;
+		errsav = errno;
+		close(sock);
+		errno = errsav;
+	}
+	return -1;
+}
+
 /**
  * Opens a UNIX server socket then manage a server based on it
  */
