+++ main/gridinit_internals.h
@@ -80,8 +80,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # define CFG_KEY_INHERIT "inherit_env"
 #endif
 
-int __open_unix_server(const char *path);
-
 int __open_unix_client(const char *path);
 
 /* Groups matching */
